requirements.txt
asgiref==3.4.1
coverage==6.0.1
django-bootstrap-pagination==1.7.1
django-widget-tweaks==1.4.8
django-with-asserts==0.0.1
Faker==8.10.3
humanize==3.12.0
libgravatar==1.0.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
pillow == 9.0.1